Fix spring-session dependency
This resolves a few issues with the Spring dependencies.

The first issue is that spring-session-data-redis is intended for non
Spring Boot applications. Spring Boot applications should use
spring-session and spring-boot-starter-redis.

The second issue is that HttpSessionListener support was not added in
spring session 1.0.3.BUILD-SNAPSHOT. It was added in the 1.1.x versions.
This means either 1.1.0.M1 or 1.1.0.BUILD-SNAPSHOT should be used.

The final issue is that in order to properly manage versions, Spring
Boot version properties should be used. This ensures all transitive
dependency versions are correctly overriden too.